Remove spec tests for now until the security setup is clarified More clean-up
