+++ x-pack/plugin/eql/qa/security/src/javaRestTest/java/org/elasticsearch/xpack/eql/EqlStatsIT.java
@@ -6,8 +6,27 @@
 
 package org.elasticsearch.xpack.eql;
 
+import org.elasticsearch.client.Request;
+import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.test.eql.stats.EqlUsageRestTestCase;
+
+import java.io.IOException;
+
+import static org.elasticsearch.xpack.eql.SecurityUtils.secureClientSettings;
 
 public class EqlStatsIT extends EqlUsageRestTestCase {
 
+    /**
+     * All tests run as a superuser but use <code>es-security-runas-user</code> to become a less privileged user.
+     */
+    @Override
+    protected Settings restClientSettings() {
+        return secureClientSettings();
+    }
+
+    @Override
+    protected void runRequest(Request request) throws IOException {
+        SecurityUtils.setRunAsHeader(request,"test-admin");
+        super.runRequest(request);
+    }
 }
